added benchmark helper to make assessing benchmarks easier
