+++ app/Filament/Resources/ResultResource.php
@@ -306,10 +306,6 @@ class ResultResource extends Resource
                     ->formatStateUsing(function ($state) {
                         return number_format((float) $state, 2, '.', '').' %';
                     }),
-                Tables\Columns\IconColumn::make('healthy')
-                    ->alignCenter()
-                    ->boolean()
-                    ->toggleable(isToggledHiddenByDefault: true),
                 Tables\Columns\TextColumn::make('status')
                     ->badge()
                     ->toggleable()
@@ -334,9 +330,6 @@ class ResultResource extends Resource
                     ->alignment(Alignment::End),
             ])
             ->filters([
-                Tables\Filters\TernaryFilter::make('healthy')
-                    ->nullable(),
-
                 Tables\Filters\SelectFilter::make('ip_address')
                     ->label('IP address')
                     ->multiple()
